Fix bug in two checksum-generated scripts
ce-sum.sh and ctcs-sum.sh are used to generate checksum. The checksum-
result files need to be put in as same directory as the scripts'.

Before, the both scripts outputted the checksum-result files directly
to as same directory as theirs. On the other hand, this directory will
also be scanned to caculate checksum. So, you can see the instantly
outputted files will affect the checsum result.

To fix this, now the both scripts output the checksum-result files to
/tmp/ first, then after everything is done, move checksum-result files
back to the directory where they are.