Add option -nosalt into openssl command line
Option -nosalt can make sure the encrypted result will not be changed if
the source data and the password are not changed.

The result file outputted by update-ver.sh is encrypted with file
'VERSION'. We can decrypt it, view it and then re-encrypt it with
'VERSION' again. The re-encrypted file should not be different from
the original one.