Fix bug that $TIME doesn't work
When TIME=time, Command line '$TIME ./filesum.sh' doesn't work on
several systems which don't have a command file 'time' in filesystem.

There are two 'time', one is bash keyword, another is an actual binary
command file in /bin/ or /usr/bin/. After expand variable TIME, bash
gets the value 'time', and it does not interprete this value again. That
means it does not consider the value 'time' as its keyword but the
binary command. So on the system which doesn't have command file 'time',
this command line fails.

To fix this and to use keyword 'time' rather than command 'time' always.
The command line is changed into 'time ./filesum.sh'. You can see I use
hardcode 'time' here instead of variable TIME, that can make sure the
'time' will be considered as the keyword by bash. And I use variable
TIMEFORMAT which is a builtin variable of bash to turn on/off the time
output.