Fix a small bug in makedelta2.sh
In function make_batch_file(), the 'exit' calling can not make the
script exit, because that 'exit' is in a subshell (after a pipe
separator, '|').
To fix it, let it exit with a non-zero value, 'exit 1', that means the
pipeline in which the 'exit' calling is will return error, which will
cause the script exit because I use 'set -e' at the beginning of the
script.