Fix bug in case there is no file to delete or add
Before, if there is no file to delete or add, the script will not
generate two intermediate files, $COPY_FILE_LIST and $DEL_FILE_LIST,
which will cause error for subsequent commands which read the both
intermediate files.

Now the both intermediate files will be generated unconditionally, but
empty if there is no file to delete or add. The empty file is friendly
to subsequent commands, so no need to modify them regardless the both
intermediate files will be generated or not.