Assure scripts can be executed in any directory
Assure the both 'used-for-validate' scripts can be executed in any
directory, not only in directory where they are.

Because they call another script whose path relative to them is fixed.
So to be able to execute them in any directory, I must use command
'dirname' to extract their directory first, then use the relative path
to find that script they call.