Intermediate files can be output to specified dir
Use the 3rd argument to specify dir to which the intermediate files are
output. The default is ".", the current directory where the script is
running. The other usual value is "/tmp".

If you execute this script in directory which is just the right one
being compared by this script, you'd better use the 3rd argument to
specify another dir instead of ".", the current dir, in case the
intermediate files is compared too.