Add a new old-new-dir package for testing
The new package contains hardlinks.
This package is consist of two parts, one is old-new-dir-w-hl.cpio.xz
which is similar with the original no-hardlink one, old-new-dir.cpio.xz,
another is old-new-dir-w-hl-add.tar.xz which is a supplement to the
former. The latter, the .tar.xz file, contains two symlinks which are
hard link to each other. The two symlinks are the supplement to .cpio.xz
file, because 'cpio' can not preserve hardlink of symbolic link, but
'tar' can.

Script 'makedelta2.sh' can handle hardlink well, however, not perfect.
After execute the delta-update script generated by makedelta2.sh, in the
two directories, the content of regular files, the type of special files,
the device major and minor number of device files and what the symlinks
point to all become same. In this sense, we can say the two directories
are same, but not exactly same. The relationship of some hardlink files
can not be sync'ed. This is mainly due to 'diff'. 'diff' usually doesn't
care wheter a file is a hardlink or not if its content is same with
another one. Anyway, nevertheless, if we don't care the hardlinks (i.e.
we do not expect the relationship of all hardlinks can be sync'ed) or
there is almost no hardlinks in directory which needs to be sync'ed,
then I can tell so far the script works pretty good.